+++ src/org/traccar/handler/events/FuelDropEventHandler.java
@@ -16,7 +16,6 @@
 package org.traccar.handler.events;
 
 import io.netty.channel.ChannelHandler;
-import org.traccar.BaseEventHandler;
 import org.traccar.Context;
 import org.traccar.model.Device;
 import org.traccar.model.Event;
