+++ src/org/traccar/handler/events/OverspeedEventHandler.java
@@ -20,7 +20,6 @@ import java.util.Collections;
 import java.util.Map;
 
 import io.netty.channel.ChannelHandler;
-import org.traccar.BaseEventHandler;
 import org.traccar.Context;
 import org.traccar.model.Device;
 import org.traccar.model.DeviceState;
