+++ src/org/traccar/sms/smpp/ClientSmppSessionHandler.java
@@ -18,7 +18,6 @@ package org.traccar.sms.smpp;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.traccar.handler.events.TextMessageEventHandler;
 
 import com.cloudhopper.commons.charset.CharsetUtil;
 import com.cloudhopper.smpp.SmppConstants;
