+++ src/org/traccar/sms/smpp/TextMessageEventHandler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.traccar.handler.events;
+package org.traccar.sms.smpp;
 
 import org.traccar.Context;
 import org.traccar.model.Device;
